+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Logging when a director cancels office hours or does not clock in for office hours. These will fall under logtype "director-cancellation" and "director-absent" respectively.
  - "start" property for each specialBreaks in configuration; these are queued at the start of a broadcast.
  - "end" property for each specialBreaks in configuration; these are queued when a broadcast ends.
+ - sails.helpers.songs.reloadSubcategories() which reloads RadioDJ subcategory numbers into configuration.
 
 ### Changed
  - Many of the methods used in api controllers and in models have been migrated to sails helpers.